Implement Character item details as separate screens
