+++ common/src/commonMain/kotlin/cz/frantisekmasa/wfrp_master/common/character/skills/SkillsCard.kt
@@ -17,11 +17,9 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.benasher44.uuid.Uuid
 import cz.frantisekmasa.wfrp_master.common.character.CharacterScreenModel
 import cz.frantisekmasa.wfrp_master.common.character.skills.dialog.AddBasicSkillsDialog
 import cz.frantisekmasa.wfrp_master.common.character.skills.dialog.AddSkillDialog
-import cz.frantisekmasa.wfrp_master.common.character.skills.dialog.EditSkillDialog
 import cz.frantisekmasa.wfrp_master.common.core.domain.Stats
 import cz.frantisekmasa.wfrp_master.common.core.domain.skills.Skill
 import cz.frantisekmasa.wfrp_master.common.core.shared.Resources
@@ -31,6 +29,7 @@ import cz.frantisekmasa.wfrp_master.common.core.ui.cards.CardItem
 import cz.frantisekmasa.wfrp_master.common.core.ui.cards.CardTitle
 import cz.frantisekmasa.wfrp_master.common.core.ui.flow.collectWithLifecycle
 import cz.frantisekmasa.wfrp_master.common.core.ui.menu.DropdownMenu
+import cz.frantisekmasa.wfrp_master.common.core.ui.navigation.LocalNavigationTransaction
 import cz.frantisekmasa.wfrp_master.common.core.ui.primitives.ContextMenu
 import cz.frantisekmasa.wfrp_master.common.core.ui.primitives.EmptyUI
 import cz.frantisekmasa.wfrp_master.common.localization.LocalStrings
@@ -94,23 +93,21 @@ internal fun SkillsCard(
                     size = EmptyUI.Size.Small
                 )
             } else {
-                var editedSkillId: Uuid? by rememberSaveable { mutableStateOf(null) }
+                val navigation = LocalNavigationTransaction.current
 
                 for (skill in skills) {
                     SkillItem(
                         skill,
                         characteristics,
-                        onClick = { editedSkillId = skill.id },
+                        onClick = {
+                            navigation.navigate(
+                                CharacterSkillDetailScreen(
+                                    skillsScreenModel.characterId,
+                                    skill.id,
+                                )
+                            )
+                        },
                         onRemove = { onRemove(skill) },
-                    )
-                }
-
-                editedSkillId?.let { skillId ->
-                    EditSkillDialog(
-                        screenModel = skillsScreenModel,
-                        skillId = skillId,
-                        characteristics = characteristics,
-                        onDismissRequest = { editedSkillId = null }
                     )
                 }
             }
