+++ common/src/commonMain/kotlin/cz/frantisekmasa/wfrp_master/common/character/talents/TalentsCard.kt
@@ -10,15 +10,14 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.benasher44.uuid.Uuid
 import cz.frantisekmasa.wfrp_master.common.character.talents.dialog.AddTalentDialog
-import cz.frantisekmasa.wfrp_master.common.character.talents.dialog.EditTalentDialog
 import cz.frantisekmasa.wfrp_master.common.core.domain.talents.Talent
 import cz.frantisekmasa.wfrp_master.common.core.ui.buttons.CardButton
 import cz.frantisekmasa.wfrp_master.common.core.ui.cards.CardContainer
 import cz.frantisekmasa.wfrp_master.common.core.ui.cards.CardItem
 import cz.frantisekmasa.wfrp_master.common.core.ui.cards.CardTitle
 import cz.frantisekmasa.wfrp_master.common.core.ui.flow.collectWithLifecycle
+import cz.frantisekmasa.wfrp_master.common.core.ui.navigation.LocalNavigationTransaction
 import cz.frantisekmasa.wfrp_master.common.core.ui.primitives.ContextMenu
 import cz.frantisekmasa.wfrp_master.common.localization.LocalStrings
 
@@ -37,21 +36,20 @@ internal fun TalentsCard(
 
             if (talents.isNotEmpty()) {
                 Column {
-                    var editedTalentId: Uuid? by rememberSaveable { mutableStateOf(null) }
+                    val navigation = LocalNavigationTransaction.current
 
                     for (talent in talents) {
                         TalentItem(
                             talent,
-                            onClick = { editedTalentId = talent.id },
+                            onClick = {
+                                navigation.navigate(
+                                    CharacterTalentDetailScreen(
+                                        screenModel.characterId,
+                                        talent.id,
+                                    )
+                                )
+                            },
                             onRemove = { onRemove(talent) }
-                        )
-                    }
-
-                    editedTalentId?.let {
-                        EditTalentDialog(
-                            screenModel = screenModel,
-                            talentId = it,
-                            onDismissRequest = { editedTalentId = null },
                         )
                     }
                 }
