+++ common/src/commonMain/kotlin/cz/frantisekmasa/wfrp_master/common/character/traits/TraitsCard.kt
@@ -10,15 +10,14 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.benasher44.uuid.Uuid
 import cz.frantisekmasa.wfrp_master.common.character.traits.dialog.AddTraitDialog
-import cz.frantisekmasa.wfrp_master.common.character.traits.dialog.EditTraitDialog
 import cz.frantisekmasa.wfrp_master.common.core.domain.traits.Trait
 import cz.frantisekmasa.wfrp_master.common.core.ui.buttons.CardButton
 import cz.frantisekmasa.wfrp_master.common.core.ui.cards.CardContainer
 import cz.frantisekmasa.wfrp_master.common.core.ui.cards.CardItem
 import cz.frantisekmasa.wfrp_master.common.core.ui.cards.CardTitle
 import cz.frantisekmasa.wfrp_master.common.core.ui.flow.collectWithLifecycle
+import cz.frantisekmasa.wfrp_master.common.core.ui.navigation.LocalNavigationTransaction
 import cz.frantisekmasa.wfrp_master.common.core.ui.primitives.ContextMenu
 import cz.frantisekmasa.wfrp_master.common.localization.LocalStrings
 
@@ -37,21 +36,20 @@ internal fun TraitsCard(
 
             if (traits.isNotEmpty()) {
                 Column {
-                    var editedTraitId: Uuid? by rememberSaveable { mutableStateOf(null) }
+                    val navigation = LocalNavigationTransaction.current
 
                     for (trait in traits) {
                         TraitItem(
                             trait,
-                            onClick = { editedTraitId = trait.id },
+                            onClick = {
+                                navigation.navigate(
+                                    CharacterTraitDetailScreen(
+                                        screenModel.characterId,
+                                        trait.id,
+                                    )
+                                )
+                            },
                             onRemove = { onRemove(trait) }
-                        )
-                    }
-
-                    editedTraitId?.let {
-                        EditTraitDialog(
-                            screenModel = screenModel,
-                            traitId = it,
-                            onDismissRequest = { editedTraitId = null },
                         )
                     }
                 }
