+++ common/src/commonMain/kotlin/cz/frantisekmasa/wfrp_master/common/localization/Strings.kt
@@ -431,6 +431,7 @@ data class CommonUiStrings(
     val markdownSupportedNote: String = "Markdown formatting is supported",
     val decrement: String = "Decrement value",
     val increment: String = "Increment value",
+    val itemDoesNotExist: String = "Item does not exist.",
     val itemNone: String = "None",
     val iconToggleFabMenu: String = "Toggle menu",
     val dismissTipConfirmation: String = "Do you really want to dismiss this tip?",
@@ -550,7 +551,6 @@ data class CompendiumMessageStrings(
     val rulebookImportFailed: String = "PDF import failed. Check that you provided valid rulebook PDF.",
     val jsonImportFailed: String = "JSON import failed. Check that you provided valid WFRP Master export.",
     val itemAlreadyExists: String = "Item already exists",
-    val itemDoesNotExist: String = "Item does not exist.",
     val noItems: String = "No items in compendium",
     val noItemsInCompendiumSubtextPlayer: String = "Your GM has to add them first.",
     val willReplaceExistingItem: String = "Will replace existing item",
